PoC: Use Redis for rate limiting/request counts
When running cc with multiple puma workers the in-memory approach
for rate limiting & concurrent request limiting (service brokers)
is not working anymore as the worker processes are independent.
Furthermore Puma distributes requests by load and thus there will be no
equal request distribution.

This change replaces the in-memory parts with Redis. This also
simplifies the logic in some parts as Redis keys have a TTL.
Redis is supposed to be deployed on each VM and thus no HA setup, networking etc. is required.

This is only a poc!
Following things are missing or should be investigated
- Redis connection pooling
- Thread/Worker safety when interacting with Redis
- Redis integration into capi-release
- Error handling (skip rate limiting if redis is not available)